Change to use the most populated cell _for which we have data_.
In the previous version of this code, we used the most populated cell
in the municipality without checking if we actually had data for that cell.
This commit changes the logic to use the first cell for which we have data.

The reason is that the snow model no longer always has data for the most
populated cell. We then select the second most populated cell etc., until
we find a cell with a non-NA value.

The snow model was changed last summer to align with the grid from the regular seNorge model (temperature, precipitation etc).

In the previous version of this code we took the mean of the cells intersecting with the cell from SSB, and thus avoided this problem.